Add option to only validate PR title
The previous implementation could only be configured to check the PR
title in cases where the PR had multiple commits, so single commit PRs
would still fallback to using the commit.

I've added an option (for backwards compatability) to instead always
use the PR title. Setting this to true causes the bot to ignore the
existing validateEitherPrOrSingleCommitTitle setting.

The sample yml config has been updated, with the previous
default behaviour left as-is.