Title validator supports custom description
This commit allows the end user to customise the error message that
displays when validation of PR/commit titles fails.

Users can include three replacement strings for validation type
(commit or PR), the title of the failing entity, and the regex used
for validation.

In order to support existing deployments, the default message is
as per the previous behaviour.